+++ src/pub/har_viewer.js
@@ -267,7 +267,7 @@ class Actions extends Pure_component {
         const _this = this;
         this.etask(function*(){
             this.on('uncaught', e=>_this.etask(function*(){
-                yield report_exception(e);
+                yield report_exception(e, 'har_viewer.Actions.resend');
             }));
             yield window.fetch('/api/logs_resend', {
                 method: 'POST',