+++ src/pub/proxy_edit/users_modal.js
@@ -39,7 +39,7 @@ export default class Users_modal extends Pure_component {
                 console.log('finally');
             });
             this.on('uncaught', e=>_this.etask(function*(){
-                yield report_exception(e);
+                yield report_exception(e, 'users_modal.Users_modal.load');
             }));
             console.log('loading');
             const users = yield ajax.json({url});